Redesign tracking of gases in the global atmosphere
The concept is to try to track any gases that get added or removed from
the global atmosphere. The game will skip these changes, but by
monitoring the GasMixtures they get added to/removed from, we can track
them.

The main sources of gas changes are:
- TakeAtmospherePortion: Gas is taken from all neighbour cells and then
  mixed and re-distributed. If a neighbour cell is global atmosphere,
  the gas is just created, so we deduct it from our tracked global.
- GiveAtmospheresMixInWorld: The result of atmosphere mixing, if gas is
  being distributed into a global atmosphere cell, its just deleted - so
  instead we track it here.
- LerpAtmosphere: World Atmosphere cells are designed to naturally gravitate
  towards the Global Atmosphere. We track any gas changes due to this
  mechanic.
- Deregistering a cell: When a cell gets close to global atmosphere, it
  gets removed, and any remaining gas in it deleted. This is a minor
  correction factor, but still good to track.

This entire process has in testing shown to be relatively accurate,
within single digit percentages, when gas is added or removed from the
world. However, it is not perfect.

One key complication is that when the reference global atmosphere is
changed, any active cells "desync" from that value, and will trigger
mixing to re-fill - resulting in more changes to the global atmosphere.
Due to the size of the atmosphere, this effect is miniscule, but likely
part of further inaccuracies.