Update for the Rocket update
Biggest change is the lack of saving of the atmosphere by the game,
instead we save it our selves in terraforming_atmosphere.xml

The atmosphere tracking is mostly unchanged, only adapted for game
changes.

SolarScale seems to be missing, and the temperature curve might need to
be adapted for the new celestial model.